+++ Iron_XCoder_Functions/error.h
@@ -5,8 +5,8 @@
 extern "C"{
 #endif
 
-void error_banner(const char *header, const char *text, int location_x, int location_y, int width, int height, int background_color);
-void error_notification(int x, int y, int width, int height, int color);
+void error_banner(const char *header, const char *text, uint8_t location_x, uint8_t location_y, uint8_t width, uint8_t height, uint8_t background_color);
+void error_notification(uint8_t x, uint8_t y, uint8_t width, uint8_t height, uint8_t color);
 
 #ifdef _cplusplus
 }
